Add ability to ignore errors for AssetDetails
GetAssetDetails can sometimes throw a 500 on one of the details URLs
(e.g. assessment). This change allows those errors to be skipped. In
addition, if the field `assessment` isn't present when creating the
asset object, this is now handled properly. Future changes might want to
expand this out to other fields as well.